new changes made

+++ frontend/src/pages/admin/RecipesList.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from "react";
 import axios from "axios";
 import { useNavigate } from "react-router-dom";
 import moment from "moment";
+import Swal from "sweetalert2";
 import AdminSideBar from "../../components/admin/AdminSideBar";
 import AdminNavbar from "../../components/adminNavbar/AdminNavbar";
 import Footer from "../../components/Footer/Footer";
@@ -41,7 +42,9 @@ const RecipeModal = ({ recipe, onClose }) => {
           <span className="close" onClick={handleCloseModal}>
             &times;
           </span>
-          <h2>{recipe.title}</h2>
+          <h2 style={{ paddingBottom: "20px", textDecoration: "underline" }}>
+            {recipe.title}
+          </h2>
           <img
             src={recipe.img}
             alt={recipe.title}
@@ -106,7 +109,11 @@ const RecipeModal = ({ recipe, onClose }) => {
               <div className="media-button">
                 <button onClick={handleThumbnailClick}>View Thumbnail</button>
                 {showThumbnail && (
-                  <img src={recipe.thumbnail} alt="Recipe Thumbnail" />
+                  <img
+                    src={recipe.thumbnail}
+                    alt="Recipe Thumbnail"
+                    style={{ width: "30%" }}
+                  />
                 )}
               </div>
             </div>
@@ -116,6 +123,7 @@ const RecipeModal = ({ recipe, onClose }) => {
     </div>
   );
 };
+
 const RecipesList = () => {
   const [recipes, setRecipes] = useState([]);
   const [filteredRecipes, setFilteredRecipes] = useState([]);
@@ -194,7 +202,9 @@ const RecipesList = () => {
   };
 
   const handleNextPage = () => {
-    setPageNo(pageNo + 1);
+    if (pageNo < Math.ceil(totalRecipes / 20)) {
+      setPageNo(pageNo + 1);
+    }
   };
 
   const handleSearch = (e) => {
@@ -211,7 +221,7 @@ const RecipesList = () => {
     if (
       !isNaN(pageNumber) &&
       pageNumber >= 1 &&
-      pageNumber <= Math.ceil(totalRecipes / 50)
+      pageNumber <= Math.ceil(totalRecipes / 20)
     ) {
       setPageNo(pageNumber);
     }
@@ -228,21 +238,36 @@ const RecipesList = () => {
   };
 
   const handleRemoveRecipe = async (recipeId) => {
-    try {
-      const authToken = getAuthToken();
-      await axios.delete("http://localhost:8000/recipe/deleterecipe/", {
-        headers: {
-          Authorization: `Bearer ${authToken.access}`,
-        },
-        data: {
-          recipeid: recipeId,
-        },
-      });
-      setRecipes(recipes.filter((recipe) => recipe.recipeid !== recipeId));
-    } catch (error) {
-      console.error("Error removing recipe:", error);
-      if (error.response && error.response.status === 401) {
-        await handleTokenRefresh(() => handleRemoveRecipe(recipeId));
+    const confirmed = await Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    });
+
+    if (confirmed.isConfirmed) {
+      try {
+        const authToken = getAuthToken();
+        await axios.delete("http://localhost:8000/recipe/deleterecipe/", {
+          headers: {
+            Authorization: `Bearer ${authToken.access}`,
+          },
+          data: {
+            recipeid: recipeId,
+          },
+        });
+        setRecipes(recipes.filter((recipe) => recipe.recipeid !== recipeId));
+        setTotalRecipes(totalRecipes - 1);
+
+        Swal.fire("Deleted!", "The recipe has been deleted.");
+      } catch (error) {
+        console.error("Error removing recipe:", error);
+        if (error.response && error.response.status === 401) {
+          await handleTokenRefresh(() => handleRemoveRecipe(recipeId));
+        }
       }
     }
   };
@@ -269,7 +294,7 @@ const RecipesList = () => {
                   placeholder="No"
                   min="1"
                   className="page-input"
-                  max={Math.ceil(totalRecipes / 50)}
+                  max={Math.ceil(totalRecipes / 20)}
                   style={{ width: "100px", marginRight: "5px" }}
                   onKeyDown={(e) => {
                     if (e.code === "Minus" || e.key === "-" || e.key === ".") {
@@ -344,7 +369,7 @@ const RecipesList = () => {
               Prev
             </button>
             <span>
-              {pageNo} of {Math.ceil(totalRecipes / 50)}
+              {pageNo} of {Math.ceil(totalRecipes / 20)}
             </span>
             <button
               onClick={handleNextPage}
